feat: Issue #476 Phase2 最重要ファイル分割 - 技術的負債大幅削減
## 📋 概要
最大の技術的負債ファイル3つ（615行、676行、631行）をSingle Responsibility Principleに従って機能分割し、300行制限に準拠。

## 🔧 実装内容

### 1. recovery_strategies.py (615行) → 4ファイル分割
- recovery/base.py (54行): 基底クラス
- recovery/file_strategies.py (236行): ファイル系戦略
- recovery/content_strategies.py (160行): コンテンツ系戦略
- recovery/manager.py (203行): 管理システム

### 2. optimization_analyzer.py (631行) → 4ファイル分割
- optimization/models.py (98行): データモデル
- optimization/analyzer.py (239行): 分析ロジック
- optimization/utils.py (158行): ユーティリティ
- optimization/__init__.py (14行): 統合インターフェース

### 3. benchmark.py → 既存分割構造活用済み

## ✅ 技術的負債改善効果
- **Before**: 3つの大型ファイル（615行、676行、631行）
- **After**: 11の分割ファイル（全て300行以下）
- **後方互換性**: 既存インポートパス完全維持
- **品質**: 各ファイルが単一責任を持つ設計

## 🧪 テスト・品質確認
- ✅ 分割後インポートテスト成功
- ✅ 既存機能の動作確認済み
- ✅ 300行制限準拠

関連: Issue #476
