+++ kumihan_formatter/core/performance/optimization/__init__.py
@@ -1,0 +1,14 @@
+"""
+最適化分析システム - Issue #402対応
+
+最適化効果の測定、分析、レポート生成機能。
+"""
+
+from .analyzer import OptimizationAnalyzer
+from .models import OptimizationMetrics, OptimizationReport
+
+__all__ = [
+    "OptimizationAnalyzer",
+    "OptimizationMetrics",
+    "OptimizationReport",
+]
